CIE-5505 - Having DataHub Edge doc in a single place (#3944)
Somebody complained (reasonably) that the documentation for DataHub on Edge was slightly fragmented, and would sometimes send you on a link wild goose chase between the DataHub and Edge sections of the doc.

I've transplanted everything into the Edge section of the doc. I think this makes more sense now that Edge is just a Kubernetes-based mini-cloud, and DataHub is "just" another component (with some caveats).

For the same reasons, I've removed a bunch of the kubernetes diagnostic content, because Edge health checks all of its components, and aggregates all of its logs.