Changed availability date for release candidates
In the retrospective of the 2025 release we discovered that one week of upgrade testing is not sufficient for the Cloud operations team. The time span has been increased to 2 weeks.For e.g. 2026 this results in an RC availability on the 6th of Feb, which a Friday an thus moved to 9th (Monday),.